atualizaçãozinh felas v2 sem virus 2026 atualizado gratis

+++ src/main/java/com/fernando/evento/entity/BlocoHorario.java
@@ -1,26 +1,25 @@
 package com.fernando.evento.entity;
 
-import jakarta.persistence.Id;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
 @Entity
 public class BlocoHorario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) 
-    private Long id;
+    private Integer id;
 
     private String descricao;
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
